requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
groq>=0.4.0
